+++ app/src/main/java/com/app/arcabyolimpo/data/remote/api/ArcaApi.kt
@@ -14,14 +14,13 @@ import com.app.arcabyolimpo.data.remote.dto.password.RecoverPasswordResponseDto
 import com.app.arcabyolimpo.data.remote.dto.password.UpdatePasswordDto
 import com.app.arcabyolimpo.data.remote.dto.password.UpdatePasswordResponseDto
 import com.app.arcabyolimpo.data.remote.dto.password.VerifyTokenResponseDto
-import com.app.arcabyolimpo.data.remote.dto.product.ProductDto
-import com.app.arcabyolimpo.data.remote.dto.supplies.AcquisitionDto
-import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchDto
 import com.app.arcabyolimpo.data.remote.dto.product.ProductDetailDto
+import com.app.arcabyolimpo.data.remote.dto.product.ProductDto
 import com.app.arcabyolimpo.data.remote.dto.product.ProductRegisterInfoDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchModifyDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchRegisterDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.AcquisitionDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.DeleteDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.DeleteResponseDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.DeleteSupplyBatchDto
@@ -135,7 +134,7 @@ interface ArcaApi {
 
     @POST("/supplyBatch/filter")
     suspend fun filterSupplyBatch(
-        @Body body: FilterRequestDto
+        @Body body: FilterRequestDto,
     ): List<FilteredBatchDto>
 
     @GET("supplyBatch/filter/data")
@@ -283,7 +282,6 @@ interface ArcaApi {
         @Part imagenInsumo: MultipartBody.Part?,
     )
 
-
     @GET("productBatch/")
     suspend fun getProductBatches(): List<ProductBatchDto>
 
@@ -304,10 +302,14 @@ interface ArcaApi {
     )
 
     @GET("productBatch/search")
-    suspend fun searchProductBatch(@Query("q") term: String): List<ProductBatchDto>
+    suspend fun searchProductBatch(
+        @Query("q") term: String,
+    ): List<ProductBatchDto>
 
     @POST("productBatch/filter")
-    suspend fun filterProductBatch(@Body filters: FilterDto): List<ProductBatchDto>
+    suspend fun filterProductBatch(
+        @Body filters: FilterDto,
+    ): List<ProductBatchDto>
 
     @DELETE("productBatch/{id}")
     suspend fun deleteProductBatch(
@@ -341,7 +343,9 @@ interface ArcaApi {
     ): List<ProductDto>
 
     @GET("product/{id}")
-    suspend fun getProductById(@Path("id") productId: String): ProductDto?
+    suspend fun getProductById(
+        @Path("id") productId: String,
+    ): ProductDto?
 
     @GET("product/add")
     suspend fun getProductFilters(): ProductRegisterInfoDto
@@ -378,7 +382,7 @@ interface ArcaApi {
         @Path("date") expirationDate: String,
         @Path("idInsumo") idSupply: String,
     ): List<SupplyBatchItemDto>
-        
+
     @GET("product/{idProduct}/update")
     suspend fun getProduct(
         @Path("idProduct") idProduct: String,
@@ -394,6 +398,6 @@ interface ArcaApi {
         @Part("idCategoria") idCategory: RequestBody,
         @Part("Descripcion") description: RequestBody,
         @Part("Disponible") status: RequestBody,
-        @Part image: MultipartBody.Part?
+        @Part image: MultipartBody.Part?,
     )
 }