+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/productbatches/productBatchRegister/ProductBatchRegisterUiState.kt
@@ -13,8 +13,16 @@ import com.app.arcabyolimpo.presentation.screens.productbatches.model.ProductBat
 data class ProductBatchRegisterUiState(
     val idProducto: String = "",
     val precioVenta: String = "",
-    val cantidadProducida: String = "",
+    val cantidadProducida: String = "0",
     val fechaCaducidad: String = "",
     val fechaRealizacion: String = "",
     val error: String? = null,
+    val isIdProductoError: Boolean = false,
+    val isPrecioVentaError: Boolean = false,
+    val isCantidadProducidaError: Boolean = false,
+    val isFechaRealizacionError: Boolean = false,
+    val isFechaCaducidadError: Boolean = false,
+    val productIds: List<String> = emptyList(),
+    val names: List<String> = emptyList(),
+    val isLoadingProducts: Boolean = false,
 )